Server: Add debug API endpoint for generating a large number of changes
This commit modifies `populateDatabase.ts` to:
- Be runnable from the api/debug (dev-mode-only) endpoint.
- Create shares, in addition to users and items.
