+++ packages/server/src/tools/debug/clearDatabase.ts
@@ -1,0 +1,6 @@
+import { DbConnection, dropTables, migrateLatest } from '../../db';
+
+export default async function clearDatabase(db: DbConnection) {
+	await dropTables(db);
+	await migrateLatest(db);
+}
